refactor: CardGameReadyPage 내부 코드를 MiniGameReadyPage로 옮김
CardGameReadyPage에서만 쓰이는 컴포넌트가 아니라서, MiniGameReadyPage로 옮겨서 전체 게임에서도 이 컴포넌트를 사용할 수 있도록 함
