--- frontend/src/features/miniGame/cardGame/pages/CardGameReadyPage/CardGameReadyPage.stories.tsx
@@ -1,28 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react-webpack5';
-import CardGameReadyPage from './CardGameReadyPage';
-
-const meta = {
-  title: 'Features/MiniGame/CardGame/CardGameReadyPage',
-  component: CardGameReadyPage,
-  parameters: {
-    layout: 'centered',
-  },
-  tags: ['autodocs'],
-} satisfies Meta<typeof CardGameReadyPage>;
-
-export default meta;
-
-type Story = StoryObj<typeof CardGameReadyPage>;
-
-export const Default: Story = {
-  render: () => {
-    return (
-      <div style={{ width: '100vw', height: '100vh' }}>
-        <CardGameReadyPage />
-      </div>
-    );
-  },
-  parameters: {
-    layout: 'centered',
-  },
-};
