--- frontend/src/features/miniGame/cardGame/pages/CardGameReadyPage/CardGameReadyPage.styled.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Image = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-`;
