+++ frontend/src/features/miniGame/pages/MiniGameReadyPage/MiniGameReadyPage.styled.ts
@@ -1,30 +1,7 @@
 import styled from '@emotion/styled';
 
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+export const SlideImage = styled.img`
+  width: 100%;
   height: 100%;
-  padding: 0 20px;
-`;
-
-export const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 16px;
-  margin-bottom: 30px;
-`;
-
-export const DescriptionWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-`;
-
-export const Timer = styled.span`
-  font-size: 150px;
-  font-weight: 900;
-  color: white;
+  object-fit: contain;
 `;
